Move connected component searchInput to containers folder
